+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Changes since v2.18
 - Pdf generation failed for applications that had table fields with no rows. Now fixed.
 
 ### Additions
-- Pdf applications now contain license text (for inline licenses), url (for link licenses) or attachment name (for attachment licenses). (#2688)
+- Pdf applications now contain license text (for inline licenses), url (for link licenses) or attachment name (for attachment licenses), and license acceptance status for applicant and members. (#2688)
 
 ## v2.18 "Särkiniemientie" 2021-05-18